mongodb-gcp/device_simulation/requirements.txt
flask
google-api-python-client
google-cloud-iot
paho-mqtt
PyJWT
cryptography
pyyaml
google-cloud-error-reporting